+++ src/Vankosoft/AgentBundle/Form/VirtualHostForm.php
@@ -1,0 +1,41 @@
+<?php namespace Vankosoft\AgentBundle\Form;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+
+class VirtualHostForm extends AbstractType
+{
+    public function buildForm( FormBuilderInterface $builder, array $options ): void
+    {
+        $builder
+            ->add( 'virtualHost', TextType::class, [
+                'label'                 => 'vs_agent.form.virtual_host.virtual_host',
+                'translation_domain'    => 'VSAgentBundle',
+                'required'              => true
+            ])
+            
+            ->add( 'apacheService', TextType::class, [
+                'label'                 => 'vs_agent.form.virtual_host.apache_service',
+                'translation_domain'    => 'VSAgentBundle',
+                'required'              => true
+            ])
+        ;
+    }
+    
+    public function configureOptions( OptionsResolver $resolver ): void
+    {
+        parent::configureOptions( $resolver );
+        
+        $resolver->setDefaults([
+            'csrf_protection'   => false,
+        ]);
+    }
+    
+    public function getName()
+    {
+        return 'vs_agent_virtual_host';
+    }
+}