+++ src/Vankosoft/AgentBundle/Resources/themes/default/assets/js/pages/actions-index.js
@@ -1,0 +1,78 @@
+// bin/console fos:js-routing:dump --format=json --target=public/shared_assets/js/fos_js_routes_admin.json
+import { VsPath } from '@@/js/includes/fos_js_routes.js';
+import { VsSpinnerShow, VsSpinnerHide } from '@@/js/includes/vs_spinner.js';
+
+$( function()
+{
+    $( '#btnShowVirtualHostForm' ).on( 'click', function( e )
+    {
+        /** Bootstrap 5 Modal Toggle */
+        const myModal = new bootstrap.Modal( '#brokeVirtualHostModal', {
+            keyboard: false
+        });
+        myModal.show( $( '#brokeVirtualHostModal' ).get( 0 ) );
+    });
+
+    $( '#btnShowBrokenVirtualHost' ).on( 'click', function( e )
+    {
+        let formData    = new FormData();
+        formData.set( "virtualHost", $( '#virtual_host_form_virtualHost' ).val() );
+        
+        VsSpinnerShow();
+        $.ajax({
+            type: 'GET',
+            url: VsPath( 'vs_agent_actions_show_broken_virtual_host' ),
+            data: formData,
+            processData: false,
+            contentType: false,
+            type: 'POST',
+            success: function ( response ) {
+                VsSpinnerHide();
+                
+                if ( response.status == 'ok' ) {
+                    alert( response.data.filecontents );
+                    document.location = document.location;
+                } else {
+                    alert( 'RESPONSE ERROR!!!' );
+                }
+            }, 
+            error: function(XMLHttpRequest, textStatus, errorThrown) {
+                VsSpinnerHide();
+                
+                alert( 'FATAL ERROR!!!' );
+            }
+        });
+    });
+    
+	$( '#btnBrokeVirtualHost' ).on( 'click', function ( e )
+    {
+        let formData    = new FormData();
+        formData.set( "virtualHost", $( '#virtual_host_form_virtualHost' ).val() );
+        formData.set( "apacheService", $( '#virtual_host_form_virtualHost' ).val() );
+        
+        VsSpinnerShow();
+        $.ajax({
+            type: 'GET',
+            url: VsPath( 'vs_agent_actions_broke_virtual_host' ),
+            data: formData,
+            processData: false,
+            contentType: false,
+            type: 'POST',
+            success: function ( response ) {
+                VsSpinnerHide();
+                
+                if ( response.status == 'ok' ) {
+                    alert( response.data.filecontents );
+                    document.location = document.location;
+                } else {
+                    alert( 'RESPONSE ERROR!!!' );
+                }
+            }, 
+            error: function(XMLHttpRequest, textStatus, errorThrown) {
+                VsSpinnerHide();
+                
+                alert( 'FATAL ERROR!!!' );
+            }
+        });
+    });
+});